fix: many warnings fixed

+++ x64barebones/Kernel/syscalls/sys_alloc.c
@@ -3,5 +3,5 @@
 // en vez de retornar la el void *, nos pasa una posicion de memoria y se lo ponemos ahi
 // para no tener que cambiar el tipo de retorno de todas las syscalls
 uint64_t sys_alloc(uint64_t len) {
-	return mm_malloc(len);
+	return (uint64_t) mm_malloc(len);
 }
